+++ ivy/functional/backends/tensorflow/experimental/norms.py
@@ -69,7 +69,7 @@ def instance_norm(
 
 
 def lp_normalize(
-    x: Union[tf.Tensor, tf.Variable], /, *, p: int = 2, axis: int = None, out=None
+    x: Union[tf.Tensor, tf.Variable], /, *, p: float = 2, axis: int = None, out=None
 ) -> tf.Tensor:
     denorm = tf.norm(x, ord=p, axis=axis, keepdims=True)
     denorm = tf.math.maximum(denorm, 1e-12)
